requirements.txt
requests
langchain
langchain-google-genai
streamlit
python-dotenv
PyPDF2

-e .
